+++ src/App.tsx
@@ -1,12 +1,12 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import Home from "./pages/homepage";
-import Login from "./pages/login";
-import SignUp from "./pages/signup";
-import UserView from "./pages/userprofile";
-import ProjectDetailsPage from "./pages/ProjectDetailsPage";
-import UploadProject from "./pages/UploadProject";
-import UploadExperienceReport from "./pages/UploadExperienceReport";
-import Error from "./pages/404";
+import Home from "./pages/homePage/homePage";
+import Login from "./pages/loginPage/loginPage";
+import SignUp from "./pages/signupPage/signupPage";
+import UserView from "./pages/userProfilePage/userProfilePage";
+import ProjectDetailsPage from "./pages/ProjectDetailsPage/projectDetailsPage";
+import UploadProject from "./pages/uploadProject/uploadProject";
+import UploadExperienceReport from "./pages/uploadExperienceReport/uploadExperienceReport";
+import Error from "./pages/404Page/404Page";
 import { AdapterDayjs } from "@mui/x-date-pickers/AdapterDayjs";
 import { LocalizationProvider } from "@mui/x-date-pickers";
 import { createTheme, ThemeProvider } from "@mui/material/styles";
