+++ src/pages/404Page/404Page.css
@@ -1,4 +1,4 @@
-@import "main.css";
+@import "../../styles/main.css";
 #error{
     background-color: rgba(159, 196, 163,0);
     top: 25%;
