+++ src/pages/signupPage/signupPage.css
@@ -1,0 +1,78 @@
+@import "../../styles/main.css";
+
+main {
+  margin: auto;
+}
+
+.wholeLogin {
+  display: flex;
+  align-items: center;
+  min-height: 100vh;
+}
+
+.loginBanner {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  align-items: center;
+  flex-basis: 40%;
+  height: 100vh;
+  background: var(--tertiaryGreen);
+  box-sizing: border-box;
+}
+
+.loginBannerImage {
+  position: relative;
+  width: 100%;
+  max-height: 120px;
+  object-fit: cover;
+  object-position: center;
+}
+
+.loginText,
+.signupText {
+  color: var(--mainGreen);
+  font-size: min(3vw, 1.7rem);
+  font-weight: var(--headerFontWeight);
+  align-items: center;
+  margin-inline: 30px;
+}
+
+.loginText p {
+  margin: 0;
+}
+
+.imgLogo {
+  position: relative;
+  width: 25vw;
+  max-width: 240px;
+  padding-bottom: 100px;
+}
+
+.loginPart {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  height: 100vh;
+  flex-basis: 60%;
+}
+
+.signupRedirect {
+  width: 90%;
+  text-align: end;
+}
+
+.signupLink {
+  font-weight: var(--headerFontWeight);
+  text-decoration: none;
+  color: var(--mainGreen);
+}
+
+.signupLink:hover {
+  text-decoration: underline;
+}
+
+.loginHeader {
+  font-family: var(--mainFontFamily) !important;
+  font-weight: var(--headerFontWeight) !important;
+}
