+++ src/pages/uploadExperienceReport/uploadExperienceReport.css
@@ -1,4 +1,4 @@
-@import "main.css";
+@import "../../styles/main.css";
 
 .outline{
     text-align: center;
