+++ src/pages/uploadProject/uploadProject.css
@@ -1,0 +1,35 @@
+@import "../../styles/main.css";
+
+.outline{
+    text-align: center;
+
+}
+
+.title{
+    font-family: var(--mainFontFamily);
+    color: var(--mainGreen);
+    font-size: 4em;
+}
+
+.group {
+    display:block;
+    font-size: 1.5em;
+}
+
+.inputField{
+    background-color: #E3E3E3;
+    border: none;
+    min-height: 3em;
+    border-radius: 1em;
+    width:100%;
+    padding-left: 1em;
+}
+
+.uploadProjectform{
+    display: grid;
+    gap: 4em;
+    grid-template-columns: repeat(2, 1fr);
+    margin: 2em;
+    margin-left: 10em;
+    margin-right: 10em;
+}
