+++ src/pages/userProfilePage/userProfilePage.css
@@ -1,4 +1,4 @@
-@import "main.css";
+@import "../../styles/main.css";
 
 .user{
     background-color: var(--mainBackground);
